# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Distribuidora API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestItemsItem } from "./createOrderRequestItemsItem";

export interface CreateOrderRequest {
  items: CreateOrderRequestItemsItem[];
  notes?: string;
  shippingAddress?: string;
}