# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Distribuidora API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  name: string;
  description?: string;
  price: number;
  stock: number;
  sku?: string;
  imageUrl?: string;
  categoryId?: number;
  isActive?: boolean;
}