# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Distribuidora API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  stock: number;
  sku?: string | null;
  imageUrl?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  isActive: boolean;
  createdAt: string;
}